Day 9 better

+++ src/util/Util.scala
@@ -1,9 +1,7 @@
 package util
 
-import scala.collection.mutable
 import scala.io.Source
 import scala.util.Using
-import java.util.concurrent.{BlockingQueue, LinkedBlockingQueue}
 
 object Util {
   def loadDay(day: Int): String = {
@@ -16,12 +14,11 @@ object Util {
 
   def loadDayInts(day: Int): List[Int] = loadDayLines(day).map(_.toInt)
 
-  def loadDayProgram(day: Int): Map[Int, BigInt] =
+  def loadDayProgram(day: Int): Map[Long, Long] =
     Util.loadDay(day)
     .split(",")
-    .map { x => BigInt(x) }
+    .map { x => x.toLong }
     .zipWithIndex
-    .map { x => (x._2, x._1) }.toMap
-
+    .map { x => (x._2.toLong, x._1) }.toMap
 }
